Cleanup crate root by moving headless and window to modules
